requirements.txt
Django>=3.2
django-filter>=25.1
djangorestframework>=3.16.0
josepy>=2.1.0
python-jose>=3.3.0
requests>=2.32.5